Add all remaining methods and fix dependency on HTTP::Headers::Fast:
Some tests were failing because, while we didn't demand
HTTP::Headers::Fast anymore, we still actually used it in tests.

This resulted a problem with missing a bunch of methods provided by
them. Those methods were added. We also added helper methods to both
HTTP::Headers and HTTP::Headers::Fast which weren't available in
HTTP::XSHeaders, so that's fixed it.

Lastly, the documentation was fixed to:

* Include all of these methods.
* Direct the user to read their documentation from the originating
  modules.
* Put them in the right order (except for new() which goes at the
  top).